safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEImageElement-svgdom-preserveAspectRatio-prop-expected.txt

CONSOLE MESSAGE: line 35: TypeError: 'undefined' is not an object (evaluating 'image.preserveAspectRatio.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAspectRatio' property of the SVGFEImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL image.preserveAspectRatio.baseVal.align should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'image.preserveAspectRatio.baseVal')
FAIL image.preserveAspectRatio.baseVal.meetOrSlice should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'image.preserveAspectRatio.baseVal')
